+++ NEWS.md
@@ -3,6 +3,10 @@
 # labelr 0.1.6
 * `ssort()` code is streamlined (and no longer issues a warning when it recycles the descending argument). 
 
+* added fact.to.char (defaults to FALSE) and irreg.to.na (defaults to FALSE) arguments to `as_base_data_frame()`.
+
+* removed extraneous code from `transfer_labs()`.
+
 * Corrected argument specification error in example `flab()` and `slab()` calls in the Introduction vignette (call was supplying argument "Female" when it should have supplied "F") . Vignette call syntax error was introduced in v1.0.5 and was limited to the syntax of the specific example calls in the vignette. The functions themselves were not at fault and are, thus, unchanged.
 
 # labelr 0.1.5